Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    float num;
-    scanf("%f", &num);
-    printf("You entered: %.4f\n", num);
+    double num;
+    scanf("%.1f", &num);
+    printf("You entered: %.1f\n", num);
     return 0;
 }
